requirements.txt
Django==4.1.3
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.3.0
inflection==0.5.1
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
psycopg2==2.9.7
googletrans==4.0.0-rc1
openai=0.28.0
openai-whisper==20230314
ffmpeg-python==0.2.0
gTTS==2.3.2
